+++ src/main/java/org/tvl/tvlooker/persistence/repository/DirectorRepository.java
@@ -1,7 +1,20 @@
 package org.tvl.tvlooker.persistence.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 import org.tvl.tvlooker.domain.model.entity.Director;
 
-interface DirectorRepository extends JpaRepository<Director, Long> {
+import java.util.Optional;
+
+/**
+ * Repository for Director entity persistence operations.
+ */
+@Repository
+public interface DirectorRepository extends JpaRepository<Director, Long> {
+
+    /**
+     * Finds a director by their TMDB ID.
+     */
+    Optional<Director> findByTmdbId(Long tmdbId);
 }
+
