+++ src/main/java/org/tvl/tvlooker/persistence/repository/ItemRepository.java
@@ -1,7 +1,26 @@
 package org.tvl.tvlooker.persistence.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 import org.tvl.tvlooker.domain.model.entity.Item;
+import org.tvl.tvlooker.domain.model.enums.TmdbType;
 
-interface ItemRepository extends JpaRepository<Item, Long> {
+import java.util.Optional;
+
+/**
+ * Repository for Item entity persistence operations.
+ */
+@Repository
+public interface ItemRepository extends JpaRepository<Item, Long> {
+
+    /**
+     * Finds an item by its TMDB ID and type (MOVIE or TV).
+     */
+    Optional<Item> findByTmdbIdAndTmdbType(Long tmdbId, TmdbType tmdbType);
+
+    /**
+     * Checks if an item exists by its TMDB ID and type.
+     */
+    boolean existsByTmdbIdAndTmdbType(Long tmdbId, TmdbType tmdbType);
 }
+
