+++ src/main/resources/application.properties
@@ -62,15 +62,15 @@ recommendation.weights.matrix-factorization=0.10
 # TMDB API Authentication (API Read Access Token v4)
 # Get yours at: https://www.themoviedb.org/settings/api
 # Set via environment variable for security: TMDB_API_KEY=your-token
-tmdb.api.key=${TMDB_API_KEY:}
-tmdb.api.base-url=https://api.themoviedb.org/3
+tmdb.api.key=eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIzZWMwMWI4ZDk5MjVjZmNlYzFjNjM3NzI3NzIwZTQ0MCIsIm5iZiI6MTc3MzUyODUyMy41MDcsInN1YiI6IjY5YjVlNWNiNjg0ZDhiM2Y0MTNiOTI1MCIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.SV60Bbl4yBIEOOrBHg9aLGrzg_5jRX_WSoG3GiXFsIQ
+tmdb.api.base-url=https://api.themoviedb.org/3/
 
 # Language for TMDB responses (titles, overviews, etc.)
 tmdb.api.language=es-MX
 
 # Collector: initial bulk load configuration
 # Set to true to run the collector when the application starts
-tmdb.collector.run-on-startup=false
+tmdb.collector.run-on-startup=true
 # Max pages to fetch per category (20 items/page)
 tmdb.collector.max-pages=50
 # Delay between API calls in ms (~25 req/s, safely under TMDB's 40 req/s limit)
@@ -80,7 +80,7 @@ tmdb.collector.request-delay-ms=40
 # Whether the synchronizer is enabled
 tmdb.sync.enabled=true
 # Sync interval in ms (default: 24 hours)
-tmdb.sync.interval-ms=86400000
+tmdb.sync.interval-ms=604800000
 # Initial delay before first sync in ms (default: 60 seconds)
 tmdb.sync.initial-delay-ms=60000
 # Pages of popular items to check for new content during sync
