+++ src/test/resources/application-test.properties
@@ -27,9 +27,9 @@ spring.jpa.show-sql=true
 spring.h2.console.enabled=false
 
 # TMDB API (disabled for tests)
-tmdb.api.key=test-disabled
+tmdb.api.key=eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIzZWMwMWI4ZDk5MjVjZmNlYzFjNjM3NzI3NzIwZTQ0MCIsIm5iZiI6MTc3MzUyODUyMy41MDcsInN1YiI6IjY5YjVlNWNiNjg0ZDhiM2Y0MTNiOTI1MCIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.SV60Bbl4yBIEOOrBHg9aLGrzg_5jRX_WSoG3GiXFsIQ
 tmdb.api.base-url=https://api.themoviedb.org/3
 tmdb.api.language=en-US
-tmdb.collector.run-on-startup=false
+tmdb.collector.run-on-startup=true
+tmdb.collector.max-pages=1
 tmdb.sync.enabled=false
-
